Fixes from midi game implementation.

+++ sop_src/utils/buf_alloc.cpp
@@ -56,7 +56,7 @@ BufAlloc::ba_t *BufAlloc::MallocWords(const size_t numWords) {
   const size_t availwords = AvailWords();
   if (numWords <= availwords) {
     rval = &mpWordBufAry[mBufIdx];
-    mBufIdx += numWords;
+    mBufIdx += (int)numWords;
   }
   LOG_ASSERT(rval);
   return rval;
